Added orders table to main
