+++ package.json
@@ -8,6 +8,7 @@
     "@heroicons/react": "^1.0.6",
     "@mui/icons-material": "^5.6.1",
     "@mui/material": "^5.6.1",
+    "@mui/x-data-grid": "^5.12.1",
     "@types/react": "latest",
     "@types/react-dom": "latest",
     "firebase": "^9.8.1",
